% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FSAWs-internals.R
\name{fitPlot}
\alias{fitPlot}
\title{Fitted model plots.}
\usage{
fitPlot(object, ...)
}
\arguments{
\item{object}{An object (see details).}

\item{\dots}{Other arguments to be passed to the plot functions.}
}
\value{
None. However, a fitted-line plot is produced.
}
\description{
A generic function for constructing a fitted model plot.
}
\details{
Specifics described relative to specific object (e.g., see \code{\link{emp}}).
}
