% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FroeseWs.R
\name{FroeseWs}
\alias{FroeseWs}
\alias{plot.FroeseWs}
\alias{coef.FroeseWs}
\title{Computes the standard weight equation using the methods described in Froese (2006).}
\usage{
FroeseWs(log.a, b)

\method{coef}{FroeseWs}(object, ...)

\method{plot}{FroeseWs}(
  x,
  min,
  max,
  what = c("both", "raw", "log"),
  col.pop = "rainbow",
  lwd.pop = 1,
  lty.pop = 1,
  order.pop = TRUE,
  col.Ws = "black",
  lwd.Ws = 3,
  lty.Ws = 1,
  ...
)
}
\arguments{
\item{log.a}{A numeric vector that contains the \eqn{log_{10}(a)} values for the population of length-weight regression equations.}

\item{b}{A numeric vector that contains the \eqn{b} values for the population of length-weight regression equations}

\item{\dots}{Additional arguments for methods.}

\item{x, object}{An object saved from the \code{FrowseWs()} call (i.e., of class \code{Froese}).}

\item{min}{A number that indicates the midpoint value of the smallest X-mm length to model.}

\item{max}{A number that indicates the midpoint value of the largest X-mm length category.}

\item{what}{A string that indicates the type of plot to produce. See details.}

\item{col.pop}{A string that indicates the type of color or palette to use for the population of length-weight regression lines. See details.}

\item{lwd.pop}{A numeric that indicates the width of the line to use for the population of length-weight regression lines.}

\item{lty.pop}{A numeric that indicates the type of line to use for the population of length-weight regression lines.}

\item{order.pop}{A logical that indicates whether the populations should be plotted from the smallest to largest weight in the initial length category. See details.}

\item{col.Ws}{A string that indicates the type of color to use for the standard length-weight regression line.}

\item{lwd.Ws}{A numeric that indicates the width of the line to use for the standard length-weight regression line.}

\item{lty.Ws}{A numeric that indicates the type of line to use for the standard length-weight regression line.}
}
\value{
A list is returned with the following items:
\itemize{
  \item \code{log.a} is a numeric vector of the observed \eqn{log_{10}(a)} values sent in the \code{log.a} argument.
  \item \code{b} is a numeric vector of the observed \eqn{b} values sent in the \code{b} argument.
  \item \code{gm.a} is a numeric that contains the geometric mean of the \eqn{a} parameter. This is simply the back-transformed mean \eqn{log_{10}(a)} value -- i.e., \eqn{10^{log_{10}(a)}}.
  \item \code{mn.b} is the arithmetic mean of the \eqn{b} parameter. 
  \code{mn.log.a} is the arithmetic mean of \eqn{log_{10}(a)}.
}
}
\description{
Computes the standard weight equation using the geometric mean of \eqn{a} and the mean of \eqn{b} from weight-length regression equations as described in Froese (2006).
}
\details{
The main function computes the mean of the \eqn{log_{10}(a)} and \eqn{b} values for the standard weight equation as detailed in Froese (2006). Note that  \eqn{log_{10}(a)} and \eqn{b} must be from the regression of \eqn{log_{10}(W)} on \eqn{log_{10}(L)} where \eqn{W} is measured in grams and \eqn{L} is the total length measured in mm.

The \code{what} argument in the \code{plot} method can be set to \code{"both"}, \code{"log"}, or \code{"raw"}. The \code{"raw"} plot shows lines on the length-weight scale for each population with the resultant standard weight equation superimposed in red. The \code{"log"} plot constructs a similar plot but on the \eqn{log_{10}(weight)}-\eqn{log_{10}(length)} scale. The \code{"both"} option produces both plots side-by-side. If the \code{col.pop} argument is one of \code{"rainbow"}, \code{"heat"}, \code{"topo"}, \code{"terrain"}, \code{"cm"}, \code{"default"}, or \code{"grey"} and \code{order.pop=TRUE} then the populations plotted should form a general color gradient from smallest to largest weight in the initial length category. This will make it easier to identify populations that \dQuote{cross over} other populations.

\code{coef} returns the geometric mean of \eqn{a} and the mean of \eqn{b} to serve as the standard weight equation as described in Froese (2006).
}
\examples{
#See examples in RuffeWs.

}
\references{
Froese, R. 2006. Cube law, condition factor and weight-length relationships: history, meta-analysis and recommendations. Journal of Applied Ichthyology 22:241-253.
}
\seealso{
\code{\link{rlp}}, \code{\link{emp}}, and \code{\link{wsValidate}}
}
\keyword{hplot}
\keyword{manip}
