\docType{package}
\name{FSAWs}
\alias{FSAWs}
\alias{FSAWs-package}
\title{Functions for constructing and validating standard weight (Ws) equations for fish.}
\source{
From Ogle, D.H. and I.J. Winfield.  2009.  Ruffe
length-weight relationships with a proposed standard weight
equation.  North American Journal of Fisheries Management
29:850-85.
}
\description{
Functions for constructing and validating standard weight
(Ws) equations for fish.  These functions were created for
Ogle and Winfield (2009) but have been used in several
other papers (see references below).  These functions are
not actively maintained any more.
}

