% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paletteChoices.R
\name{paletteChoices}
\alias{paletteChoices}
\title{Provides a vector of possible color palettes.}
\usage{
paletteChoices()
}
\value{
A single string with a palette name.
}
\description{
Provides a vector of possible color palettes.
}
\details{
Generally serves as an internal function and should be called by the user.
}
\author{
Derek H. Ogle, \email{DerekOgle51@gmail.com}
}
\keyword{misc}
