% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rlp.R
\name{rlp}
\alias{rlp}
\alias{plot.rlp}
\alias{anova.rlp}
\alias{coef.rlp}
\alias{predict.rlp}
\alias{fitPlot.rlp}
\alias{summary.rlp}
\title{Computes the standard weight equation using the regression-line-percentile method.}
\usage{
rlp(log.a, b, min, max, w = 10, qtype = 8, probs = 0.75, digits = NULL)

\method{plot}{rlp}(
  x,
  what = c("both", "raw", "log"),
  col.pop = "rainbow",
  lwd.pop = 1,
  lty.pop = 1,
  order.pop = TRUE,
  col.Ws = "black",
  lwd.Ws = 3,
  lty.Ws = 1,
  ...
)

\method{coef}{rlp}(object, ...)

\method{summary}{rlp}(object, ...)

\method{predict}{rlp}(object, ...)

\method{anova}{rlp}(object, ...)

\method{fitPlot}{rlp}(
  object,
  pch = 16,
  col.pt = "black",
  col.Ws = "red",
  lwd.Ws = 3,
  lty.Ws = 1,
  xlab = "log10(midpt Length)",
  ylab = paste0("log10(", 100 * object$prob, " Percentile of Predicted Weight)"),
  main = "RLP Equation Fit",
  ...
)
}
\arguments{
\item{log.a}{A numeric vector that contains the \eqn{log_{10}(a)} values for the population of length-weight regression equations.}

\item{b}{A numeric vector that contains the \eqn{b} values for the population of length-weight regression equations}

\item{min}{A number that indicates the midpoint value of the smallest X-mm length category.}

\item{max}{A number that indicates the midpoint value of the largest X-mm length category.}

\item{w}{A number that indicates the widths for which to create length categories.}

\item{qtype}{Type of quantile method to use. See description of types of quantile calculation methods in \code{\link{quantile}}.}

\item{probs}{A number that indicates the probability of the quantile. Must be between 0 and 1.}

\item{digits}{Number of digits to round predicted weights. If \code{NULL} (default), no rounding will be used.}

\item{x, object}{An object saved from the \code{rlp} call (i.e., of class \code{rlp}).}

\item{what}{A string that indicates the type of plot to produce. See details.}

\item{col.pop}{A string that indicates the type of color or palette to use for the population of length-weight regression lines. See details.}

\item{lwd.pop}{A numeric that indicates the width of the line to use for the population of length-weight regression lines.}

\item{lty.pop}{A numeric that indicates the type of line to use for the population of length-weight regression lines.}

\item{order.pop}{A logical that indicates whether the populations should be plotted from the smallest to largest weight in the initial length category. See details.}

\item{col.Ws}{A string that indicates the type of color to use for the standard length-weight regression line.}

\item{lwd.Ws}{A numeric that indicates the width of the line to use for the standard length-weight regression line.}

\item{lty.Ws}{A numeric that indicates the type of line to use for the standard length-weight regression line.}

\item{\dots}{Additional arguments for methods.}

\item{pch}{A single numeric that indicates what plotting character codes should be used for the points in \code{fitPlot}.}

\item{col.pt}{A string used to indicate the color of the plotted points.}

\item{xlab}{A label for the x-axis of \code{fitPlot}.}

\item{ylab}{A label for the y-axis of \code{fitPlot}.}

\item{main}{A label for the main title of \code{fitPlot}.}
}
\value{
The main function returns a list with the following items:
\itemize{
  \item \code{log.a} is a numeric vector of the observed \eqn{log_{10}(a)} values sent in the \code{log.a} argument.
  \item \code{b} is a numeric vector of the observed \eqn{b} values sent in the \code{b} argument.
  \item \code{data.pred} is a matrix of the predicted weight at length for all populations.
  \item \code{data.reg} contains a data frame with the \code{prob}th quartile of predicted weights and the midpoint lengths.
  \item \code{Ws} is an \code{lm} object that contains the results of the regression of \eqn{log_{10}(wq)} on \eqn{log_{10}(midpoint length)}.
}
}
\description{
Computes the standard weight equation using the regression-line-percentile (RLP) method when given \eqn{log_{10}(a)} and \eqn{b} values from (log-transformed) length-weight regressions fit to several populations.
}
\details{
The main function follows the steps of the regression-line-percentile (RLP) method detailed in Murphy \emph{et al.} (1990). In summary, the given \eqn{log_{10}(a)} and \eqn{b} values are used to predict a weight at the midpoint of each length class defined by \code{w} (e.g., 10 mm length classes if \code{w=10}) for each population; the predicted weight at the \code{prob}th percentile (wq) across all populations is then identified; and a linear regression equation is fit to the \eqn{log_{10}(wq)} and \eqn{log_{10}(midpoint length)} data.

Note that \eqn{log_{10}(a)} and \eqn{b} must be from the regression of \eqn{log_{10}(W)} on \eqn{log_{10}(L)} where \eqn{W} is measured in grams and \eqn{L} is the length (traditionally total length) measured in mm.

It appears that Murphy \emph{et al.} (1990) used \code{qtype=6} in their SAS program.

The \code{what} argument in the \code{plot} method can be set to \code{"both"}, \code{"log"}, or \code{"raw"}. The \code{"raw"} plot shows lines on the length-weight scale for each population with the resultant standard weight equation superimposed in red. The \code{"log"} plot constructs a similar plot but on the \eqn{log_{10}(weight)}-\eqn{log_{10}(length)} scale. The \code{"both"} option produces both plots side-by-side. If the \code{col.pop} argument is one of \code{"rainbow"}, \code{"heat"}, \code{"topo"}, \code{"terrain"}, \code{"cm"}, \code{"default"}, or \code{"grey"} and \code{order.pop=TRUE} then the populations plotted should form a general color gradient from smallest to largest weight in the initial length category. This will make it easier to identify populations that \dQuote{cross over} other populations.

\code{fitPlot} shows the log-transformed linear regression result; i.e., fitted line superimposed on the log-transformed \code{prob}the percentile predicted weights versus log-transformed midpoint length category value. The examples show how to make a corresponding residual plot.
 
\code{coef} returns \eqn{log_{10}(a)} and \eqn{b} values for the resultant standard weight equation. Similarly, \code{summary}, \code{anova}, and \code{predict} returns the typical results for the linear regression model used to create the standard weight equation.
}
\examples{
## Recreate Murphy et al. (1990) results for Largemouth Bass
# min and max lengths were 152 and 816
lmb.rlp <- rlp(LMBassWs$log.a,LMBassWs$b,155,815,qtype=6)
coef(lmb.rlp)
# compare to log.a=-5.379 and b=3.221

## Examples of the other extractor functions
summary(lmb.rlp)
anova(lmb.rlp)
predict(lmb.rlp)
10^predict(lmb.rlp,data.frame(logmidpt=log10(c(200,400))))

plot(lmb.rlp)
fitPlot(lmb.rlp)
# a residual plot for the linear regression
plot(lmb.rlp$Ws$residuals~lmb.rlp$Ws$fitted.values,pch=19)
abline(h=0,lty=3)

}
\references{
Murphy, B.R., M.L. Brown, and T.A. Springer. 1990. Evaluation of the relative weight (Wr) index, with new applications to walleye. North American Journal of Fisheries Management, 10:85-97.
}
\seealso{
\code{\link{emp}}, \code{\link{FroeseWs}}, and \code{\link{wsValidate}}; and \code{quantile} in \pkg{stats}
}
\author{
Derek H. Ogle, \email{DerekOgle51@gmail.com}
}
\keyword{hplot}
\keyword{manip}
